Improve EduQueueThread and fix errors from previous merge
