+++ CMakeLists.txt
@@ -29,7 +29,6 @@ add_library(Sts1CobcSw_Utility OBJECT)
 add_library(Sts1CobcSw_Periphery OBJECT)
 add_program(HelloDummy)
 add_program(EduProgramQueueThread)
-add_program(WIP)
 
 # All targets must only add the top level project directory as PUBLIC include directory
 all_targets_include_directories("$<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}>")
