+++ Sts1CobcSw/Utility/Time.hpp
@@ -1,11 +1,11 @@
 #include <rodos_no_using_namespace.h>
 
-namespace sts1cobcsw::util
+namespace sts1cobcsw::utility
 {
 //! Number of nanoseconds between 1st January 1970 and 1st January 2000
-constexpr auto rodosUnixOffset = 946'684'800 * RODOS::SECONDS;
+constexpr auto rodosUnixOffsetNanoseconds = 946'684'800 * RODOS::SECONDS;
 //! Number of seconds between 1st January 1970 and 1st January 2000
-constexpr auto rodosUnixOffsetSeconds = 946'684'800;
+constexpr auto rodosUnixOffsetDelay = 946'684'800;
 
 //! @brief Print utc system time in human readable format
 void PrintTime();
@@ -16,7 +16,16 @@ void PrintTime();
 {
     auto rodosTime = static_cast<int64_t>(unixTime);
     rodosTime = rodosTime * RODOS::SECONDS;
-    rodosTime = rodosTime - util::rodosUnixOffset;
+    rodosTime = rodosTime - rodosUnixOffsetNanoseconds;
     return rodosTime;
 }
+
+[[nodiscard]] inline auto GetUnixUtc() {
+
+    auto systemUTC = RODOS::sysTime.getUTC();
+    systemUTC += rodosUnixOffsetNanoseconds;
+    auto unixUtc = systemUTC / RODOS::SECONDS;
+    return static_cast<int32_t>(unixUtc);
+}
+
 }
